Move existing stories in to component directory
